refactor(editor): extract BufferLifecycle module
Move LSP buffer lifecycle (open, changed, save, kill) and Git
buffer lifecycle (open, changed, tracking) from the Editor GenServer
into BufferLifecycle. Includes after_save hook firing.

Extracted functions:
- lsp_buffer_opened, lsp_buffer_changed, lsp_after_command,
  lsp_after_save, lsp_after_kill
- git_buffer_opened, git_buffer_changed, start_git_buffer,
  git_buffer_update

Editor.ex dropped from 1,440 to 1,333 lines (-107 lines).
Removed unused GitBuffer alias.

All 3,888 tests pass. Dialyzer clean. Credo clean.

Refs: #307
